Refactor ModelActor termination
TerminateModelActor is now forwarded to the ModelActor itself, and the ModelActor handles and responds to it.

Also now it requires ActorMetadata in the constructor.
This needs an additional QueryDB access in the DebugRoute, to get the metadata of the actor.
This infrastructure is perhaps a first step towards enforce that getEvents can only done with appropriate membership