Use ActorMetadata in LocalRouter and ModelActor constructor
Note: ActorMetadata is serialized and used as the name part of the ActorPath for the ModelActor to send the message to.
When the ModelActor is instantiated, it parses the name into a new instance of the ActorMetadata.
This approach is used to prepare for cluster support (in which Props.create is invoked by Pekko instead of by us)

Also re-ordered the fields inside Case.java a bit

Note: LocalRouter does not use URLEncoded name in the ActorPath, but in Clustered setup, pekko runs a UTF8 encoder on the name (unclear when and why at this stage).

Generalized encoding and decoding URL parts in multiple places with the utility that we already had